requirements.txt
cryptography==38.0.4
numpy==1.23.5
opencv_python==4.6.0.66
pycryptodome==3.16.0